Fill default values when creating resource by gohan_db_create
Currently, if some properties of input data do not exist, default values
defined in schema file are not filled. In this time, mysql default values
are filled with mysql, and server error occurs with sqlite3. The big
problem here is the difference of behavior between mysql and sqlite3.

This fix enables default values defined in schema are filled in using
gohan_db_create.