gohan_raw_http: string response body type
Type []byte is converted to array of numbers in otto, which results
in obfuscated logs, requires manual conversion to String type
in order to work with response body and is generally a bad idea.

In this commit, Gohan automatically converts arrived data to string
and adds missing documentation for gohan_raw_http.