Crash on sync/snmp listener goroutine panic
If sync processing goroutine crashes for any reason, we should not
just log panic and keep on going. In such case, Gohan is rendered
unusable, but it doesn't go down completely, so it's not detectable.

Another problem is that in case of failure in any goroutines mentioned
above logging only error message is not sufficient for debugging.

This commit changes above behavior so in case of listener goroutine
panic, it will log whole stacktrace and crash whole Gohan process.