don't allow empty id during resource creation
Gohan WebUI sends empty string when form field is unfilled.
When it's saved to database in unchanged form, it would  cause
sync thread to attempt write to resource directory instead of file
and blocking all processing.

This commit changes resource creation behavior, so that it treats ""
(empty string) just the same as undefined for id property,
causing new UUID to be generated in it's place