Add support for migrating gohan database.
The support is provided based on goose library, 'pressly' fork. Given
migration scripts provided by developers it can migrate gohan database
to the newer versions, as well as to the previous versions (reverting).
Additionally, while gohan is starting up, database version is always
checked and if mismatched with exising database version, it rejects
to run and manual intervention is needed from administrator.