ESI-11891 Reject gohan startup when there is a migration pending
This commit prevents from running gohan when there is a pending migration
(which means that gohan and its extensions are in a specific version)
and the database used in gohan is in a previous version.
This check only works when no_init is set to true (ie. in production
environments).
In that case, operator must manually execute migration to the new
version making safety db dumps and other safety checks.