Small fixes to resource_management_tests.go.
Added BeforeEach() {} to be sure that we use adminAuth for "As an
admin" tests. That way the reader doesn't have to discover that fact
by checking all the lines above.

Broken a line that was a tad too long. That way the important
information (id and domain_id) is easily visible.

Changed expectations for tenant_id/domain_id filling tests to expect
results of auth.TenantID() and auth.DomainID() (which is what's
actually filled in) instead of some constants defined at the beginning
of the file.