BL-267: addressed one review comment.
Added a new test to show that tenant_id is in fact filled in and saved
to the database. It's simply not present in response made by
CreateResource because it's blacklisted for create.

Modified server_test.go, because adding an "action: read" policy to
test schema made GET on "/_all" return that schema.