[BL-267] Handle blacklisted tenant_id when creating resource (#689)
* BL-267: handle blacklisted tenant_id when creating resource.

We would always fail to create a resource when tenant_id is
blacklisted in a matching policy. This was due to the fact that we
filled the tenant_id field if it was not provided and was also defined
in the schema. That way we always matched the blacklist filter.

This fix is implemented by spliting Policy.Check method into two new
methods: CheckAccess and CheckPropertiesFilter. CheckAccess receives
the filled in tenant_id/domain_id as it requires that info to perform
authorization. CheckPropertiesFilter receives all the original
parameters without those additionally filled in. So if tenant_id was
not provided, it will not match against the blacklist/hidden filter.
Updating resource was not performing any tenant_id/domain_id filling,
so it still used the Check method (which is now implemented using the
extracted methods) as it leads to simpler code. And to smaller diff.

Added HasPropertyId method to Schema as a simple wrapper of
GetPropertyById, to simplify those callsites that do not need to
access the property, just to check its existence. Added tests for both
while at it.

* Small fixes to resource_management_tests.go.

Added BeforeEach() {} to be sure that we use adminAuth for "As an
admin" tests. That way the reader doesn't have to discover that fact
by checking all the lines above.

Broken a line that was a tad too long. That way the important
information (id and domain_id) is easily visible.

Changed expectations for tenant_id/domain_id filling tests to expect
results of auth.TenantID() and auth.DomainID() (which is what's
actually filled in) instead of some constants defined at the beginning
of the file.

* BL-267: addressed one review comment.

Added a new test to show that tenant_id is in fact filled in and saved
to the database. It's simply not present in response made by
CreateResource because it's blacklisted for create.

Modified server_test.go, because adding an "action: read" policy to
test schema made GET on "/_all" return that schema.