Improved test stability + removed unused code reported by go test 1.11
Fixed one error (declared but not used) reported by `go test` when using
go 1.11. In 1.11 `go test` runs (a subset of?) `go vet` automatically.

I was receiving a ton of "database is locked" errors when running `go
test ./...`. That's because all packages used `test.db` as name of
sqlite database. Modified server and otto packages to use dedicated db
files in order to avoid conflicts. With this change syncer tests are
the only ones that fail occasionally.