Fix assertion in OpenAPI v3 test
WithTransform() was used incorrectly - it should be inside To(), not Expect().
Due to this, assertion was constructed in such a way that it always passed,
regardless of what the code under test returned.

WithTransform() was not consistent with the rest of the code which called
transformJSON() directly, therefore it was removed.